feat: ready to remove api/auth/callback/apple
