+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "iSunFA",
-  "version": "0.8.5+137",
+  "version": "0.8.5+138",
   "private": false,
   "scripts": {
     "dev": "next dev",
