+++ src/pages/api/auth/callback/apple.ts
@@ -1,59 +1,8 @@
 import type { NextApiRequest, NextApiResponse } from 'next';
-import jwt from 'jsonwebtoken';
-import { generateAppleClientSecret } from '@/lib/utils/apple_auth';
+import { handleAppleOAuth } from '@/lib/utils/apple_auth';
 import { ISUNFA_ROUTE } from '@/constants/url';
 import { loggerError } from '@/lib/utils/logger_back';
 import { handleSignInSession } from '@/lib/utils/signIn';
-
-export async function handleAppleOAuth(code: string) {
-  if (!process.env.APPLE_CLIENT_ID || !process.env.APPLE_REDIRECT_URI) {
-    throw new Error('Apple OAuth environment variables are missing.');
-  }
-
-  // Info: (20241127 - tzuhan) 生成 Apple Client Secret
-  const clientSecret = generateAppleClientSecret();
-
-  // Info: (20241127 - tzuhan) 使用 fetch 發送請求到 Apple 的 /auth/token 端點
-  const tokenResponse = await fetch('https://appleid.apple.com/auth/token', {
-    method: 'POST',
-    headers: { 'Content-Type': 'application/x-www-form-urlencoded' },
-    body: new URLSearchParams({
-      client_id: process.env.APPLE_CLIENT_ID!,
-      client_secret: clientSecret,
-      code,
-      grant_type: 'authorization_code',
-      redirect_uri: process.env.APPLE_REDIRECT_URI!,
-    }),
-  });
-
-  if (!tokenResponse.ok) {
-    throw new Error(`Token exchange failed: ${tokenResponse.statusText}`);
-  }
-
-  const tokenData = await tokenResponse.json();
-  const { id_token: idToken } = tokenData;
-
-  // Info: (20241127 - tzuhan) 解碼 id_token 提取用戶數據
-  const decoded = jwt.decode(idToken) as {
-    email?: string;
-    sub: string;
-    name?: string;
-    image?: string;
-  };
-  const account = {
-    provider: 'apple',
-    providerAccountId: decoded?.sub,
-    type: 'oauth',
-  };
-  const user = {
-    id: decoded?.sub,
-    email: decoded?.email,
-    name: decoded?.name,
-    image: decoded?.image,
-  };
-
-  return { user, account };
-}
 
 export default async function handler(req: NextApiRequest, res: NextApiResponse) {
   if (req.method !== 'POST') {
